added compose.yaml for using postgres, refactored item and user package, create schema for user and item
