+++ compose.yaml
@@ -1,0 +1,24 @@
+services:
+  db:
+    image: postgres:16.1
+    container_name: postgres_shareit
+    ports:
+      - "7432:5432"
+    volumes:
+      - ./volumes/postgres:/var/lib/postgresql/data/
+    environment:
+      - POSTGRES_DB=shareit
+      - POSTGRES_USER=dbuser
+      - POSTGRES_PASSWORD=12345
+    healthcheck:
+      test: pg_isready -q -d $$POSTGRES_DB -U $$POSTGRES_USER
+      timeout: 5s
+      interval: 5s
+      retries: 10
+
+#  db-init:
+#    image: postgres:16.1
+#    container_name: db-init
+#    depends_on:
+#      db:
+#        condition: service_healthy
