+++ src/main/java/ru/practicum/shareit/item/ItemMapper.java
@@ -1,7 +1,4 @@
-package ru.practicum.shareit.item.mapper;
-
-import ru.practicum.shareit.item.dto.ItemDto;
-import ru.practicum.shareit.item.model.Item;
+package ru.practicum.shareit.item;
 
 public class ItemMapper {
 
@@ -21,7 +18,7 @@ public class ItemMapper {
                 .description(itemDto.getDescription())
                 .available(itemDto.getAvailable())
                 .owner(userId)
-                .request(null)
+                //.request(null)
                 .build();
     }
 }
