+++ src/main/java/ru/practicum/shareit/user/User.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.user.model;
+package ru.practicum.shareit.user;
 
 import jakarta.validation.constraints.Email;
 import lombok.Builder;
