+++ src/main/java/ru/practicum/shareit/user/UserMapper.java
@@ -1,7 +1,4 @@
-package ru.practicum.shareit.user.mapper;
-
-import ru.practicum.shareit.user.dto.UserDto;
-import ru.practicum.shareit.user.model.User;
+package ru.practicum.shareit.user;
 
 public class UserMapper {
