+++ src/main/java/ru/practicum/shareit/user/UserStorageImpl.java
@@ -1,9 +1,7 @@
-package ru.practicum.shareit.user.dao;
+package ru.practicum.shareit.user;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.user.dto.UserDto;
-import ru.practicum.shareit.user.model.User;
 
 import java.util.Collection;
 import java.util.HashMap;
