+++ src/main/resources/schema.sql
@@ -1,0 +1,14 @@
+CREATE TABLE IF NOT EXISTS users (
+    id BIGINT PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    name VARCHAR(255) NOT NULL,
+    email VARCHAR(512) NOT NULL UNIQUE
+    );
+
+CREATE TABLE IF NOT EXISTS items (
+    id BIGINT PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY NOT NULL,
+    name VARCHAR(255) NOT NULL,
+    description VARCHAR(512) NOT NULL,
+    available BOOLEAN,
+    user_id BIGINT,
+    FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE
+    );
